# Updated all logging to be lazy-loaded