The current logging isn't lazy-loaded. This means that all statements are evaluated even with logging disabled. This can have pretty a measurable impact on performance.