@nosammai 2022-06-18
Testing with the provided benchmark, this makes a huge difference in performance:

Before fix:
```
❯ python3.8 benches/large_resource_set.py TagAssetBenchmark
Filter    : resource["IamInstanceProfile"]["Arn"].matches("(.*)(?=Enterprise-Reserved-CloudCustodian.*)") && (resource["Tags"].filter(x, x["Key"] == "ASSET")[0]["Value"] != "CLOUDCUSTODIAN" && resource["Tags"].filter(x, x["Key"] == "ASSET")[0]["Value"] != "CLOUDCORESERVICES" && present(resource["Tags"].filter(x, x["Key"] == "ASSET")[0]["Value"]))
Resources : 1,000
Total Time: 19,423.9 ms
Range : 13.0 ms - 172.3 ms
Mean  : 19.19 ms
Median: 15.52 ms
```

After Fix:
```
❯ python3.8 benches/large_resource_set.py TagAssetBenchmark
Filter    : resource["IamInstanceProfile"]["Arn"].matches("(.*)(?=Enterprise-Reserved-CloudCustodian.*)") && (resource["Tags"].filter(x, x["Key"] == "ASSET")[0]["Value"] != "CLOUDCUSTODIAN" && resource["Tags"].filter(x, x["Key"] == "ASSET")[0]["Value"] != "CLOUDCORESERVICES" && present(resource["Tags"].filter(x, x["Key"] == "ASSET")[0]["Value"]))
Resources : 1,000
Total Time: 2,690.9 ms
Range : 2.0 ms - 8.9 ms
Mean  : 2.60 ms
Median: 2.45 ms
```