@eandersson 2022-09-02
> looks like one test failure on test_trace_decorator to address.

Ops. Missed that one. Fixed.